Release the OCaml lock during IO
This releases OCaml's runtime lock whenever we call C functions that do
IO. It should allow users of this to run ocaml-freetds in parallel, or
when using Async/Lwt.

I had to re-arrange some code to avoid calling any OCaml functions
while the lock was released.

We use wrappers around the release/acquire functions to ensure that we
don't try to acquire or release the lock twice, which could potentially
happen in the msg/error callbacks.